Feat(#77) : 메인에서 사용하는 타입 따로 분리
