+++ src/types/types.ts
@@ -29,42 +29,3 @@ export type MusicInfoType = {
   release: string;
   thumbnail: string;
 };
-
-export type GenreMusicInfo = {
-  artist: string;
-  genre: number;
-  lyrics: string;
-  musicId: string;
-  musicSource: string;
-  musicTitle: string;
-  release: string;
-  thumbnail: string;
-};
-
-export type MusicPreference = {
-  ballad: number;
-  dance: number;
-  hiphop: number;
-  rnb: number;
-  rock: number;
-};
-
-export type UserChar = {
-  userChar: {
-    gender?: boolean;
-    age?: number;
-    mbti?: number;
-  };
-};
-
-export type PlaylistCurrent = {
-  currentId: string;
-  userId: string;
-  currentMusicIds: string[];
-};
-
-export type Banner = {
-  adId: string;
-  userId: string;
-  imageUrl: string[];
-};
